fix: move SYMLINKS_SUPPORTED out of conftest.py to dodge a module-name clash
Broke CI on both platforms in the previous commit. tests/conftest.py
and tests/test_skills/conftest.py are both named "conftest", and
neither tests/ nor its subdirectories carry __init__.py, so pytest
imports both under the same bare module name — whichever loads first
wins the "conftest" slot in sys.modules for the rest of the session.
`from conftest import SYMLINKS_SUPPORTED` in the three symlink-guarded
test files worked when run in isolation (only the root conftest.py
was ever loaded), but collecting the suite alongside tests/test_skills
resolved "conftest" to that directory's conftest.py instead — which
has no SYMLINKS_SUPPORTED — and pytest aborted collection entirely
with an ImportError, failing the whole CI job on both ubuntu-latest
and windows-latest, not just the Windows job the symlink guards were
meant to fix.

Moved the probe to tests/_symlink_support.py, a name unique across the
whole tests/ tree, so there is nothing for it to collide with. Verified
by reproducing the exact failure first (`pytest tests/test_skills/
tests/test_scheduler/test_script_jobs.py` errored with the reported
ImportError before this commit) and confirming it collects and skips
cleanly after.
