+++ tests/conftest.py
@@ -18,29 +18,3 @@ os.environ.setdefault("AGENTOS_TURN_CALL_LOG", "0")
 # is cold. Default tests must stay offline; tests that exercise the refresh
 # opt back in with monkeypatch.setenv.
 os.environ.setdefault("AGENTOS_OPENCAP_LIVE_PRICING", "0")
-
-
-def _symlinks_supported() -> bool:
-    """Probe whether this process can actually create filesystem symlinks.
-
-    Windows requires SeCreateSymbolicLinkPrivilege — Developer Mode enabled,
-    or an elevated process — to call os.symlink(); without it every symlink
-    test fails identically with OSError: [WinError 1314]. Detected with a
-    real create-in-tempdir probe rather than a bare `sys.platform == "win32"`
-    check: Windows with Developer Mode on (or running elevated) supports
-    symlinks fine, and gating on OS name alone would skip tests that could
-    actually run there.
-    """
-    try:
-        with tempfile.TemporaryDirectory() as tmp:
-            target = Path(tmp) / "target"
-            target.write_text("x", encoding="utf-8")
-            (Path(tmp) / "link").symlink_to(target)
-        return True
-    except OSError:
-        return False
-
-
-#: Shared skip condition for tests that exercise real symlink creation.
-#: Usage: @pytest.mark.skipif(not SYMLINKS_SUPPORTED, reason="...")
-SYMLINKS_SUPPORTED = _symlinks_supported()
